+++ src/View/Components/FormCheckbox.php
@@ -13,7 +13,7 @@ class FormCheckbox extends FormElement
     public $selected;
     public $show_error;
     public $title;
-    
+
     public function __construct(
         string $name,
         array $options = [],