+++ src/View/Components/FormDatePicker.php
@@ -15,12 +15,14 @@ class FormDatePicker extends FormElement
 
     public function __construct(
         string $name,
+        string $id = null,
         string $label = '',
         $default = null,
         bool $required = false,
         bool $showError = true
     ) {
         $this->name         = $name;
+        $this->id           = $id;
         $this->label        = $label;
         $this->required     = $required;
         $this->show_error   = $showError;