+++ src/View/Components/FormInput.php
@@ -13,6 +13,7 @@ class FormInput extends FormElement
 
     public function __construct(
         string $name,
+        string $id = null,
         string $label = '',
         string $type = 'text',
         $default = null,
@@ -20,11 +21,12 @@ class FormInput extends FormElement
         bool $showError = true
     ) {
         $this->name         = $name;
-        $this->label        = $label ?? $name;
+        $this->id           = $id;
+        $this->label        = $label;
         $this->type         = $type;
         $this->required     = $required;
         $this->show_error   = $showError;
-        
+
         $this->value        = old($name, $default);
     }
 }