+++ changelog.md
@@ -2,6 +2,18 @@
 
 All notable changes to `LaravelSpinForms` will be documented in this file.
 
+## Version 1.0.2
+
+### Added
+- Updated Input, Textarea, Select, Date Picker, Date Time Picker, Time Picker. Allowing for manual designation of id attribute
+
+
+## Version 1.0.1
+
+### Added
+- [FIX] Label was not obeying attribute designation, so not displaying class / for etc
+
+
 ## Version 1.0
 
 ### Added